changed data type in shot noise

+++ ml_intuition/data/noise.py
@@ -139,7 +139,7 @@ class Shot(BaseNoise):
         n_affected, n_bands = \
             self.get_proba(data.shape[Sample.SAMPLES_DIM], self.params.pa), \
             self.get_proba(data.shape[Sample.FEATURES_DIM], self.params.pb)
-        data = data.astype(np.float)
+        data = data.astype(np.float32)
         data = np.clip(data, a_min=0, a_max=None)
         noise = np.random.poisson(data)
         noisy_bands = np.random.choice(data.shape[Sample.FEATURES_DIM],
